+++ manga_translator/detection/__init__.py
@@ -1,2 +1,3 @@
-from .detector import Detector
+from .detector import DetectorBase
 from .onnx_detection import ONNXDetection
+from .yolo_detection import YoloDetection